抽离播放组件

+++ src/Models/Models.BiliBili/Live/LiveFeedHotArea.cs
@@ -15,19 +15,19 @@ namespace Bili.Models.BiliBili
         /// 分区Id.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "area_v2_id", Required = Required.Default)]
-        public int AreaId { get; set; }
+        public long AreaId { get; set; }
 
         /// <summary>
         /// 父分区Id.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "area_v2_parent_id", Required = Required.Default)]
-        public int ParentAreaId { get; set; }
+        public long ParentAreaId { get; set; }
 
         /// <summary>
         /// 标签类型.
         /// </summary>
         [JsonProperty(NullValueHandling = NullValueHandling.Ignore, PropertyName = "tag_type", Required = Required.Default)]
-        public int TagType { get; set; }
+        public long TagType { get; set; }
     }
 
     /// <summary>
